#819 Move proxy lookups to dedicated class implementations
